Edge.cs:
Modified Subscribe method

Traveller.cs:

Added a variable currentPath to store the current path.
Modified the Start method to subscribe the traveler to the first edge of the current path.
Added a method spawnAtStartingPoint to store the current path reference.
Added a method MoveToNextEdge to move the traveler to the next edge in the path.

Path.cs:

Modified the constructor to subscribe the traveler to the starting edge and pass the generated path to the traveler.