Improved stats structures, Added more measures
The sum of this branch can be summarised as follows:

StatisticalManager:
The majority of this branch has been focused on the statisticalManager
class. This is also a gameObject that should be present throughout the
simulation. A prefab has been made for it so that it is easily spawned.
It also acts as the defacto game manager, being responsible for the
termination of the simulation, and controller (gently) spawns of
travellers.
This has all be done in a new file called StatisticalController.cs

Data Storage:
Bespoke C# classes have been screated to store data between scenes, and
ensure that relevent information is stored together and easily
accessible. This has been done fully for paths, storing metadata such as
start time end time etc...
This has been performed in part for waypoints and edges too, although
functionality must be extended.
Accessible in StatisticalStructure.cs

Termination criteria:
Although the body of this is within StatisticalManager, elements have
had to be implemented elsewhere. The TravellerManager class has been
modified to include a traveller counter, as well as a counter in the
statisticalManager class which measured travellers with finished paths.
An arbitrary 10 finished paths was selected as the termination criteria.
This was implemented using RecieveEndTime(), a function that is called
externally from the class, whenever a waypointMover terminates. This was
also used to update some metadata in the data storage elements of the
program. Once this function had registered the final path, cross
checking with the termination criteria and no. of travellers, the update
function was informed and the next scene loaded.

Data Acquisition:
This was primarily done during the initialisation of new waypointMovers.
all important data was copied into the relevant Statistical structures
class. As preivously mentioned certain fields (e.g. endtime) had to be
updated a posteriori, once the path had been complete, along with the
attribute routeChange. A method called saveData was implemented into the
building class (responsible for spawning), which did the above, bar the
posteriori appendages. Additionally, ID attributes were implemented
across a number of objects to ensure that consistency was kept.

Data Processing / Cleaning:
Before the transfewr into the next scene, our dataset had to be cleansed
of any bad datums. Each datum had each field examined, and any anomolous
or null results were removed. This was performed in PrunePathData in
StatisticalManager.

New Scene implementation:
A new scene was created, purely to show the statistics at the end of the
Sim. It was designed with a static title, and a textBox that shows all
currently implemented stats in text format. It uses TMP_TEXT to do this.
Tabs need to be implemented so that the project can switch to different
heatmaps to show more data. The scene is called statisticsScene

Statistical Measures:
The following have been implemented:
Total Travellers: An interim measure, used while there are still bugs in
the source code.

Total Travel Time: Sum of the time that travellers have travelled.

Average Traveller Velocity: As says. Currently experiencing a bug
relating to the persistance of game objects. fix being developed.

Rate of Deviation: The proportion of times that the traveller deviates
from its original path. Fully operational

Total Pollution: Rate of pollution, based off of distance travelled and
the rate of pollution for a given vehicle type.

Error Handling:
Empty path - sometimes when a path is generated is is empty for some
reason. comments have been added appropriately, but the solution in the
meantime is when a path is checked and found to be invalid, the
finishedPaths counter in StatisticalManager is incremented, otherwise
the program would never terminate.